Softuart test candidate (#34)
* SoftUART for sending data to the Zynq for monitoring
- add softuart from tivaware utils
- update pinout (including TI Pinmux files)
- add a CLI interface to enable/disable the transmitter (on by default)

Complete list of data to be sent via soft UART
* FPGA temps
* ADC values
* power supply information
* MCU uptime and GIT version

+++ projects/cm_mcu/ADCMonitorTask.c
@@ -86,7 +86,7 @@ struct ADC_Info_t ADCs[] = {
     {ADC_CTL_TS,   "TM4C_TEMP", 1.}, // this one is special, temp in C
 };
 
-static float fADCvalues[ADC_CHANNEL_COUNT]; // ADC values in volts
+static __fp16 fADCvalues[ADC_CHANNEL_COUNT]; // ADC values in volts
 
 // read-only accessor functions for ADC names and values.
